Initial cut at migrating in the multitasking code from other branch.
