+++ Assets/editor.css
@@ -410,6 +410,6 @@ div.field[placeholderText][contenteditable=true]:empty:focus:before {
 }
 
 .ipad_body {
-    padding-left: 80px;
-    padding-right: 80px;
+    padding-left: 40px;
+    padding-right: 40px;
 }
