+++ Classes/UIView+SizeClass.h
@@ -1,0 +1,7 @@
+#import <UIKit/UIKit.h>
+
+@interface UIView (SizeClass)
+
+- (BOOL)isViewHorizontallyCompact;
+
+@end
